Move lookups into own module
